Extracting Question Answering Task out from Seq2Seq tasks
- Reason:
    - Providing a QA model from huggingface as the backbone fails to load
    - huggingface doesn't seem to treat QnA task a Seq2Seq task.

- Work
    - Made the implementations of the QuestionAnswering data and task independent from Seq2Seq.
