+++ flash/text/question_answering/data.py
@@ -11,14 +11,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import json
 from dataclasses import dataclass
-from functools import partial
+from pathlib import Path
 from typing import Any, Callable, Dict, List, Optional, Union
 
 import torch
+from pytorch_lightning.trainer.states import RunningStage
 from torch import Tensor
 
 import flash
+from flash.core.data.callback import BaseDataFetcher
 from flash.core.data.data_module import DataModule
 from flash.core.data.data_source import DataSource, DefaultDataSources
 from flash.core.data.process import Postprocess, Preprocess
@@ -28,11 +31,22 @@ from flash.text.classification.data import TextDeserializer
 
 if _TEXT_AVAILABLE:
     import datasets
-    from datasets import DatasetDict, load_dataset
+    from datasets import Dataset, DatasetDict, load_dataset
     from transformers import AutoTokenizer, default_data_collator
 
+# TODO:
+#   1) Decision related
+#       i) Do we need a custom Deserializer ?
+#       ii) How to route the {question, context, answer}_column_name values to the class constructor ?
+#           1) Through the `load_data` method.
+#           2) Keep it same and ask the user to model data according to the "question, context, answer" column names.
+#   2)  Updates needed
+#       i) Update the `QuestionAnsweringPostprocess` class according to the example
+#       ii) Actual implementation of all methods and classes.
+#       iii) `SQuADDataSource` might also need an update.
 
-class Seq2SeqDataSource(DataSource):
+
+class QuestionAnsweringDataSource(DataSource):
 
     @requires_extras("text")
     def __init__(
@@ -40,7 +54,10 @@ class Seq2SeqDataSource(DataSource):
         backbone: str,
         max_source_length: int = 128,
         max_target_length: int = 128,
-        padding: Union[str, bool] = 'max_length'
+        padding: Union[str, bool] = 'max_length',
+        question_column_name: str = "question",
+        context_column_name: str = "context",
+        answer_column_name: str = "answers"
     ):
         super().__init__()
 
@@ -50,26 +67,44 @@ class Seq2SeqDataSource(DataSource):
         self.max_target_length = max_target_length
         self.padding = padding
 
+        # Setup global pre-processing requirements
+        #   pad_on_right, doc_stride, etc. are remaining
+        self.question_column_name = question_column_name
+        self.context_column_name = context_column_name
+        self.answer_column_name = answer_column_name
+
     def _tokenize_fn(
         self,
         ex: Union[Dict[str, str], str],
         input: Optional[str] = None,
         target: Optional[str] = None,
     ) -> Callable:
-        if isinstance(ex, dict):
-            ex_input = ex[input]
-            ex_target = ex[target] if target else None
-        else:
-            ex_input = ex
-            ex_target = None
 
-        return self.tokenizer.prepare_seq2seq_batch(
-            src_texts=ex_input,
-            tgt_texts=ex_target,
-            max_length=self.max_source_length,
-            max_target_length=self.max_target_length,
-            padding=self.padding,
-        )
+        # Setup the tokenizing call here
+        #   tokenized_samples = self.tokenizer(...params)
+        tokenized_samples = self.tokenizer()
+
+        stage = self._running_stage.value
+
+        if stage == RunningStage.TRAINING:
+            # Preprocess function for training
+            self._prepare_train_features(tokenized_samples)
+            pass
+
+        if self._running_stage.evaluating() or stage == RunningStage.PREDICTING:
+            # Preprocess function for eval or predict
+            self._prepare_val_features(tokenized_samples)
+            pass
+
+        return tokenized_samples
+
+    def _prepare_train_features(self, tokenized_samples: Any):
+        # Example labelling step
+        pass
+
+    def _prepare_val_features(self, tokenized_samples: Any):
+        # Relevant step
+        pass
 
     def __getstate__(self):  # TODO: Find out why this is being pickled
         state = self.__dict__.copy()
@@ -81,7 +116,7 @@ class Seq2SeqDataSource(DataSource):
         self.tokenizer = AutoTokenizer.from_pretrained(self.backbone, use_fast=True)
 
 
-class Seq2SeqFileDataSource(Seq2SeqDataSource):
+class QuestionAnsweringFileDataSource(QuestionAnsweringDataSource):
 
     def __init__(
         self,
@@ -129,7 +164,7 @@ class Seq2SeqFileDataSource(Seq2SeqDataSource):
             else:
                 dataset_dict = load_dataset(self.filetype, data_files=data_files)
 
-        dataset_dict = dataset_dict.map(partial(self._tokenize_fn, input=input, target=target), batched=True)
+        dataset_dict = dataset_dict.map(self._tokenize_fn, batched=True)
         dataset_dict.set_format(columns=columns)
         return dataset_dict[stage]
 
@@ -146,7 +181,7 @@ class Seq2SeqFileDataSource(Seq2SeqDataSource):
         self.tokenizer = AutoTokenizer.from_pretrained(self.backbone, use_fast=True)
 
 
-class Seq2SeqCSVDataSource(Seq2SeqFileDataSource):
+class QuestionAnsweringCSVDataSource(QuestionAnsweringFileDataSource):
 
     def __init__(
         self,
@@ -173,7 +208,7 @@ class Seq2SeqCSVDataSource(Seq2SeqFileDataSource):
         self.tokenizer = AutoTokenizer.from_pretrained(self.backbone, use_fast=True)
 
 
-class Seq2SeqJSONDataSource(Seq2SeqFileDataSource):
+class QuestionAnsweringJSONDataSource(QuestionAnsweringFileDataSource):
 
     def __init__(
         self,
@@ -200,17 +235,21 @@ class Seq2SeqJSONDataSource(Seq2SeqFileDataSource):
         self.tokenizer = AutoTokenizer.from_pretrained(self.backbone, use_fast=True)
 
 
-class Seq2SeqSentencesDataSource(Seq2SeqDataSource):
+class QuestionAnsweringDictionaryDataSource(QuestionAnsweringDataSource):
 
-    def load_data(
-        self,
-        data: Union[str, List[str]],
-        dataset: Optional[Any] = None,
-    ) -> List[Any]:
+    def load_data(self, data: Any, columns: List[str] = None) -> 'datasets.Dataset':
+        if columns is None:
+            columns = ["input_ids", "attention_mask", "labels"]
+            if self._running_stage.value == RunningStage.PREDICTING:
+                columns.remove("labels")
 
-        if isinstance(data, str):
-            data = [data]
-        return [self._tokenize_fn(s) for s in data]
+        stage = self._running_stage.value
+
+        dataset_dict = DatasetDict({stage: Dataset.from_dict(data)})
+        dataset_dict = dataset_dict.map(self._tokenize_fn, batched=True)
+
+        dataset_dict.set_format(columns=columns)
+        return dataset_dict[stage]
 
     def __getstate__(self):  # TODO: Find out why this is being pickled
         state = self.__dict__.copy()
@@ -222,25 +261,67 @@ class Seq2SeqSentencesDataSource(Seq2SeqDataSource):
         self.tokenizer = AutoTokenizer.from_pretrained(self.backbone, use_fast=True)
 
 
+class SQuADDataSource(QuestionAnsweringDataSource):
+
+    def load_data(self, data: str, dataset: Optional[Any] = None) -> 'datasets.Dataset':
+        stage = self._running_stage.value
+
+        file_path = data
+
+        path = Path(file_path)
+        with open(path, 'rb') as f:
+            squad_v_2_dict = json.load(f)
+
+        contexts = []
+        questions = []
+        answers = []
+        for topic in squad_v_2_dict['data']:
+            for comprehension in topic['paragraphs']:
+                context = comprehension['context']
+                for q_a_pair in comprehension['qas']:
+                    question = q_a_pair['question']
+                    for answer in q_a_pair['answers']:
+                        answer_text = answer['text']
+
+                        contexts.append(context)
+                        questions.append(question)
+                        answers.append(answer_text)
+
+        dataset_dict = DatasetDict({
+            stage: Dataset.from_dict({
+                "context": contexts,
+                "question": questions,
+                "answer": answers
+            })
+        })
+
+        dataset_dict = dataset_dict.map(self._tokenize_fn, batched=True)
+
+        return dataset_dict[stage]
+
+
 @dataclass(unsafe_hash=True, frozen=True)
-class Seq2SeqBackboneState(ProcessState):
-    """The ``Seq2SeqBackboneState`` stores the backbone in use by the
-    :class:`~flash.text.seq2seq.core.data.Seq2SeqPreprocess`
+class QuestionAnsweringBackboneState(ProcessState):
+    """The ``QuestionAnsweringBackboneState`` stores the backbone in use by the
+    :class:`~flash.text.question_answering.data.QuestionAnsweringPreprocess`
     """
 
     backbone: str
 
 
-class Seq2SeqPreprocess(Preprocess):
+class QuestionAnsweringPreprocess(Preprocess):
 
     @requires_extras("text")
     def __init__(
         self,
+        question_column_name: Optional[str] = None,
+        context_column_name: Optional[str] = None,
+        answer_column_name: Optional[str] = None,
         train_transform: Optional[Dict[str, Callable]] = None,
         val_transform: Optional[Dict[str, Callable]] = None,
         test_transform: Optional[Dict[str, Callable]] = None,
         predict_transform: Optional[Dict[str, Callable]] = None,
-        backbone: str = "sshleifer/tiny-mbart",
+        backbone: str = "t5-small",
         max_source_length: int = 128,
         max_target_length: int = 128,
         padding: Union[str, bool] = 'max_length'
@@ -256,30 +337,45 @@ class Seq2SeqPreprocess(Preprocess):
             test_transform=test_transform,
             predict_transform=predict_transform,
             data_sources={
-                DefaultDataSources.CSV: Seq2SeqCSVDataSource(
+                DefaultDataSources.CSV: QuestionAnsweringCSVDataSource(
                     self.backbone,
                     max_source_length=max_source_length,
                     max_target_length=max_target_length,
                     padding=padding,
+                    question_column_name=question_column_name,
+                    context_column_name=context_column_name,
+                    answer_column_name=answer_column_name,
                 ),
-                DefaultDataSources.JSON: Seq2SeqJSONDataSource(
+                DefaultDataSources.JSON: QuestionAnsweringJSONDataSource(
                     self.backbone,
                     max_source_length=max_source_length,
                     max_target_length=max_target_length,
                     padding=padding,
+                    question_column_name=question_column_name,
+                    context_column_name=context_column_name,
+                    answer_column_name=answer_column_name,
                 ),
-                "sentences": Seq2SeqSentencesDataSource(
+                "dict": QuestionAnsweringDictionaryDataSource(
                     self.backbone,
                     max_source_length=max_source_length,
                     max_target_length=max_target_length,
                     padding=padding,
+                    question_column_name=question_column_name,
+                    context_column_name=context_column_name,
+                    answer_column_name=answer_column_name,
                 ),
+                "squad_v2": SQuADDataSource(
+                    self.backbone,
+                    max_source_length=max_source_length,
+                    max_target_length=max_target_length,
+                    padding=padding,
+                )
             },
-            default_data_source="sentences",
+            default_data_source="dict",
             deserializer=TextDeserializer(backbone, max_source_length)
         )
 
-        self.set_state(Seq2SeqBackboneState(self.backbone))
+        self.set_state(QuestionAnsweringBackboneState(self.backbone))
 
     def get_state_dict(self) -> Dict[str, Any]:
         return {
@@ -299,7 +395,7 @@ class Seq2SeqPreprocess(Preprocess):
         return default_data_collator(samples)
 
 
-class Seq2SeqPostprocess(Postprocess):
+class QuestionAnsweringPostprocess(Postprocess):
 
     @requires_extras("text")
     def __init__(self):
@@ -310,7 +406,7 @@ class Seq2SeqPostprocess(Postprocess):
 
     @property
     def backbone(self):
-        backbone_state = self.get_state(Seq2SeqBackboneState)
+        backbone_state = self.get_state(QuestionAnsweringBackboneState)
         if backbone_state is not None:
             return backbone_state.backbone
 
@@ -336,8 +432,73 @@ class Seq2SeqPostprocess(Postprocess):
         self._tokenizer = AutoTokenizer.from_pretrained(self.backbone, use_fast=True)
 
 
-class Seq2SeqData(DataModule):
-    """Data module for Seq2Seq tasks."""
+class QuestionAnsweringData(DataModule):
+    """Data module for QuestionAnswering task."""
 
-    preprocess_cls = Seq2SeqPreprocess
-    postprocess_cls = Seq2SeqPostprocess
+    preprocess_cls = QuestionAnsweringPreprocess
+    postprocess_cls = QuestionAnsweringPostprocess
+
+    @classmethod
+    def from_squad_v2(
+        cls,
+        train_file: Optional[str] = None,
+        val_file: Optional[str] = None,
+        test_file: Optional[str] = None,
+        train_transform: Optional[Dict[str, Callable]] = None,
+        val_transform: Optional[Dict[str, Callable]] = None,
+        test_transform: Optional[Dict[str, Callable]] = None,
+        data_fetcher: Optional[BaseDataFetcher] = None,
+        preprocess: Optional[Preprocess] = None,
+        val_split: Optional[float] = None,
+        batch_size: int = 4,
+        num_workers: Optional[int] = None,
+        **preprocess_kwargs: Any,
+    ):
+        """Creates a :class:`~flash.text.question_answering.data.QuestionAnsweringData` object from the given
+        data JSON files in the SQuAD2.0 format.
+
+        Args:
+            train_file: The JSON file containing the training data.
+            val_file: The JSON file containing the validation data.
+            test_file: The JSON file containing the testing data.
+            train_transform: The dictionary of transforms to use during training which maps
+                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+            val_transform: The dictionary of transforms to use during validation which maps
+                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+            test_transform: The dictionary of transforms to use during testing which maps
+                :class:`~flash.core.data.process.Preprocess` hook names to callable transforms.
+            data_fetcher: The :class:`~flash.core.data.callback.BaseDataFetcher` to pass to the
+                :class:`~flash.core.data.data_module.DataModule`.
+            preprocess: The :class:`~flash.core.data.data.Preprocess` to pass to the
+                :class:`~flash.core.data.data_module.DataModule`. If ``None``, ``cls.preprocess_cls``
+                will be constructed and used.
+            val_split: The ``val_split`` argument to pass to the :class:`~flash.core.data.data_module.DataModule`.
+            batch_size: The ``batch_size`` argument to pass to the :class:`~flash.core.data.data_module.DataModule`.
+            num_workers: The ``num_workers`` argument to pass to the :class:`~flash.core.data.data_module.DataModule`.
+            preprocess_kwargs: Additional keyword arguments to use when constructing the preprocess. Will only be used
+                if ``preprocess = None``.
+
+        Returns:
+            The constructed data module.
+
+        Examples::
+
+            data_module = QuestionAnsweringData.from_squad_v2(
+                train_file="train.json",
+            )
+        """
+        return cls.from_data_source(
+            "squad_v2",
+            train_file,
+            val_file,
+            test_file,
+            train_transform=train_transform,
+            val_transform=val_transform,
+            test_transform=test_transform,
+            data_fetcher=data_fetcher,
+            preprocess=preprocess,
+            val_split=val_split,
+            batch_size=batch_size,
+            num_workers=num_workers,
+            **preprocess_kwargs,
+        )
